% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_to_n.R
\name{c_to_n}
\alias{c_to_n}
\alias{c_to_n.character}
\alias{c_to_n.factor}
\alias{c_to_n.numeric}
\alias{c_to_n.integer}
\alias{c_to_n.logical}
\alias{c_to_n.data.frame}
\title{Extract numeric values from character values}
\usage{
c_to_n(x, ...)

\method{c_to_n}{character}(x, ...,
  pattern_numeric = "^[+-]?(?:[0-9]+|[0-9]+\\\\.|\\\\.[0-9]+|[0-9]+\\\\.[0-9]+)(?:[eE][+-]?[0-9]+)?$",
  verbose = TRUE)

\method{c_to_n}{factor}(x, ...)

\method{c_to_n}{numeric}(x, ...)

\method{c_to_n}{integer}(x, ...)

\method{c_to_n}{logical}(x, ...)

\method{c_to_n}{data.frame}(x, ..., verbose = TRUE)
}
\arguments{
\item{x}{An object to convert from character to numeric using SDTM or ADaM
rules.}

\item{...}{Passed to other `c_to_n()` methods}

\item{pattern_numeric}{A regular expression to detect numeric values.}

\item{verbose}{Report detailed status of .}
}
\value{
For vectors, a numeric vector with `NA` at locations with non-numeric
  values.  For data.frames and similar, the same class of object with numeric
  versions of columns added.
}
\description{
Extract numeric values from character values
}
\details{
The main workhorse function is c_to_n.character, and it will
  recognize the following numeric types:

\itemize{
  \item{positive or negative (with optional sign) for all classes}
  \item{integers with an optional decimal point after the number}
  \item{floating point numbers without a number before the decimal point}
  \item{floating point numbers with a number before or after the decimal point}
  \item{scientific notation with any of the prior numeric types as the
    coefficient and an optional sign before the exponent}
}

Spaces at the beginning and end are removed with `trimws()`
}
\section{Methods (by class)}{
\itemize{
\item \code{character}: The main workhorse function of c_to_n

\item \code{factor}: For factors.

\item \code{numeric}: For numeric vectors.

\item \code{integer}: For integer vectors.

\item \code{logical}: For logical vectors (only handles all-NA case).

\item \code{data.frame}: For data.frames and similar, finds columns matching the
regular expression pattern `"^..(ST|OR).*C$"`.  data.frame method does not
replace numeric columns that already exist.
}}

