% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_conversion.R
\name{dateany_to_date}
\alias{dateany_to_date}
\title{Reformat any date format (ISO8601 character string, POSIXt, or Date) to be a
Date object}
\usage{
dateany_to_date(x)
}
\arguments{
\item{x}{An ISO8601 formatted character string, POSIXt object, or Date object}
}
\value{
A Date object
}
\description{
All values must be converted (without additional NA values created).  For
character strings, the year-month-day part is required to be in ISO 8601
format, but the (ignored) time format only requires separation by a
\code{"T"}.
}
\examples{
dateany_to_date("2022-01-02")
dateany_to_date("2022-01-02T03:04") # the time part is automatically dropped
dateany_to_date(as.Date("2022-01-02"))
dateany_to_date(as.POSIXct("2022-01-02T03:04")) # the time part is still gone
}
