% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_domain.R
\name{detect_domain}
\alias{detect_domain}
\title{Detect the SDTM or ADaM domain of a dataset (possibly via filename)}
\usage{
detect_domain(x, filename, verbose = TRUE)
}
\arguments{
\item{x}{The data}

\item{filename}{The filename associated with \code{x}}

\item{verbose}{Report domain detected and detection method?}
}
\value{
A character scalar with the name of the domain
}
\description{
Detect the SDTM or ADaM domain of a dataset (possibly via filename)
}
\concept{Domain detection}
