% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_methods.R
\name{get_domain}
\alias{get_domain}
\title{Get the domain of an Rsdtm object}
\usage{
get_domain(object)
}
\arguments{
\item{data}{An Rsdtm object}
}
\value{
The domain of the object as a character scalar
}
\description{
Get the domain of an Rsdtm object
}
