% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_sdtm.R
\name{import_sdtm}
\alias{import_sdtm}
\alias{import_sdtm_dir}
\alias{import_sdtm_file}
\title{Import an SDTM or ADaM file or directory of files}
\usage{
import_sdtm(
  path,
  extension_choice = c(".sas7bdat", ".xpt"),
  ignore_case = TRUE,
  ...
)

import_sdtm_dir(
  path,
  extension_choice = c(".sas7bdat", ".xpt"),
  ignore_case = TRUE,
  ignore_filename = c(),
  ...
)

import_sdtm_file(path, return_type = c("data.frame", "list"), ...)
}
\arguments{
\item{path}{The path to the directory or file.}

\item{extension_choice}{What file extension(s) should be searched for within
the directory?  (If more than one file with the same extension exists, the
first extension in \code{extension_choice} with a usable file will be used,
and a warning will be given for subsequent files.}

\item{ignore_case}{Passed to \code{list.files} when loading a directory.}

\item{...}{Arguments passed to \code{rio::import}}

\item{ignore_filename}{A vector of filenames not to load (case sensitive,
filename only excluding directory name)}

\item{return_type}{When loading a single file, what type of output should be
provided?}
}
\value{
If a directory or a vector of files is given in \code{path}, a list
  of data.frames; the names of the list will be the uppercase basename of the
  file without the extension.  If a single file is given in \code{path}, a
  data.frame.
}
\description{
Import an SDTM or ADaM file or directory of files
}
\details{
It is an error to import the same filename base (filename without
  path or extension) more than once.  That can happen by loading more than
  one file with the same name from different directories or multiple versions
  of the same file.
}
\section{Functions}{
\itemize{
\item \code{import_sdtm_dir()}: Load a directory of SDTM files.

\item \code{import_sdtm_file()}: Load a single SDTM file.

}}
