% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso8601_regexp.R
\name{is_ISO8601_calendar_date}
\alias{is_ISO8601_calendar_date}
\title{Determine if a string is an ISO 8601 date}
\usage{
is_ISO8601_calendar_date(x, ...)
}
\arguments{
\item{x}{A vector of character strings to test}

\item{...}{
  Arguments passed on to \code{\link[=pattern_ISO8601_calendar_date]{pattern_ISO8601_calendar_date}}
  \describe{
    \item{\code{truncated}}{Should the date be allowed to be truncated?  An integer
indicating the highest required precision (0=day is required, 1=month, and
2=year).  A value of 3 will allow an empty string to match.}
    \item{\code{allow_before_year_1583}}{Should years between 0 and 1582 be allowed
(they are only allowed in ISO 8601 with mutual agreement)}
    \item{\code{time_pattern}}{A string to add to the day for including time with date
(see \code{pattern_ISO8601_time()})}
  }}
}
\description{
Determine if a string is an ISO 8601 date
}
\seealso{
Other ISO8601 String checking: 
\code{\link{is_ISO8601_calendar_datetime}()},
\code{\link{is_ISO8601_timezone}()},
\code{\link{is_ISO8601_time}()}
}
\concept{ISO8601 String checking}
