% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso8601_regexp.R
\name{is_ISO8601_calendar_datetime}
\alias{is_ISO8601_calendar_datetime}
\title{Determine if a string matches an ISO 8601 date/time using standard calendar
date notation}
\usage{
is_ISO8601_calendar_datetime(x, ...)
}
\arguments{
\item{x}{A vector of character strings to test}

\item{...}{
  Arguments passed on to \code{\link[=pattern_ISO8601_calendar_datetime]{pattern_ISO8601_calendar_datetime}}
  \describe{
    \item{\code{truncated}}{Should the date/time be allowed to be truncated?  An integer
indicating the highest required precision (0=second is required, 1=minute,
2=hour, 3=day, 4=month, 5=year, 6=none).  A value of 6 will allow an empty
string to match.}
  }}
}
\description{
Determine if a string matches an ISO 8601 date/time using standard calendar
date notation
}
\seealso{
Other ISO8601 String checking: 
\code{\link{is_ISO8601_calendar_date}()},
\code{\link{is_ISO8601_timezone}()},
\code{\link{is_ISO8601_time}()}
}
\concept{ISO8601 String checking}
