% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso8601_regexp.R
\name{is_ISO8601_week_datetime}
\alias{is_ISO8601_week_datetime}
\title{Determine if a string matches an ISO 8601 date/time using week date notation}
\usage{
is_ISO8601_week_datetime(x, ...)
}
\arguments{
\item{x}{A vector of character strings to test}

\item{...}{
  Arguments passed on to \code{\link[=pattern_ISO8601_week_datetime]{pattern_ISO8601_week_datetime}}
  \describe{
    \item{\code{truncated}}{Should the date/time be allowed to be truncated?  An integer
indicating the highest required precision (0=second is required, 1=minute,
2=hour, 3=day, 4=week, 5=year, 6=none).  A value of 6 will allow an empty
string to match.}
  }}
}
\description{
Determine if a string matches an ISO 8601 date/time using week date notation
}
\seealso{
Other ISO8601 String checking: 
\code{\link{is_ISO8601_calendar_datetime}()},
\code{\link{is_ISO8601_calendar_date}()},
\code{\link{is_ISO8601_ordinal_datetime}()},
\code{\link{is_ISO8601_timezone}()},
\code{\link{is_ISO8601_time}()}

Other Date management and conversion: 
\code{\link{dateany_to_date}()},
\code{\link{generate_dtc}()},
\code{\link{is_ISO8601_calendar_datetime}()},
\code{\link{is_ISO8601_calendar_date}()},
\code{\link{is_ISO8601_ordinal_datetime}()},
\code{\link{is_ISO8601_timezone}()},
\code{\link{is_ISO8601_time}()},
\code{\link{make_dy}()},
\code{\link{pattern_ISO8601_any_datetime}()},
\code{\link{pattern_ISO8601_any_date}()},
\code{\link{pattern_ISO8601_calendar_datetime}()},
\code{\link{pattern_ISO8601_calendar_date}()},
\code{\link{pattern_ISO8601_calendar_year}()},
\code{\link{pattern_ISO8601_ordinal_datetime}()},
\code{\link{pattern_ISO8601_ordinal_date}()},
\code{\link{pattern_ISO8601_timezone}()},
\code{\link{pattern_ISO8601_time}()},
\code{\link{pattern_ISO8601_week_datetime}()},
\code{\link{pattern_ISO8601_week_date}()},
\code{\link{sdtm_dtc_to_datetime}()},
\code{\link{sdtm_first_dose}()},
\code{\link{sdtm_time_actual}()}
}
\concept{Date management and conversion}
\concept{ISO8601 String checking}
