% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_creation.R
\name{is_valid_sdtm_version}
\alias{is_valid_sdtm_version}
\title{Check if the SDTM version, implemetation guide (IG), and IG version are valid}
\usage{
is_valid_sdtm_version(sdtm_version)
}
\arguments{
\item{sdtm_version}{Version of the SDTM standard as a character string or `base::package_version()`}

\item{ig_name}{Name of the implementation guide (IG) as a character string}

\item{ig_version}{Version of the implementation guide as a character string or `base::package_version()`}
}
\value{
TRUE if it is valid, FALSE if it is not valid
}
\description{
Check if the SDTM version, implemetation guide (IG), and IG version are valid
}
