% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_conversion.R
\name{make_dy}
\alias{make_dy}
\title{Calculate an SDTM --DY value from a date and a reference date}
\usage{
make_dy(dates, refdt)
}
\arguments{
\item{dates}{The dates for calculation}

\item{refdt}{The dates for reference}
}
\value{
An integer vector with days since the reference date according to
  SDTM calculation rules:  If \code{dates < refdt}, the difference in days;
  if \code{dates >= refdt}, the difference in days plus one.
}
\description{
Calculate an SDTM --DY value from a date and a reference date
}
\details{
\code{dates} and \code{refdt} is passed through
  \code{dateany_to_date()} to ensure that they are R Date class objects.
}
\examples{
make_dy(c("2022-01-02", "2022-01-03", "2022-01-04"), "2022-01-03")
}
