% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_param_label.R
\name{make_param_label}
\alias{make_param_label}
\title{Generate a parameter label combining specimen type, parameter name, and units}
\usage{
make_param_label(
  spec,
  param,
  unit,
  expect_single = TRUE,
  allow_missing_spec = TRUE,
  allow_missing_unit = TRUE
)
}
\arguments{
\item{spec}{Specimen type (typically, "blood", "plasma", etc.)}

\item{param}{Parameter name (the thing that was measured)}

\item{unit}{Units of measure}

\item{expect_single}{Should a single parameter label be generated (or may it
be a vector)?}

\item{allow_missing_spec, allow_missing_unit}{Are the \code{spec} or
\code{unit} parameters allowed to be NA?}
}
\value{
A character vector that is equivalent to \code{sprintf("%s %s (%s)",
  spec, param, unit)}, but accounting for the fact that spec and unit may be
  missing.
}
\description{
Generate a parameter label combining specimen type, parameter name, and units
}
