% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_domain.R
\name{make_sdtm_domain_from_filename}
\alias{make_sdtm_domain_from_filename}
\title{Simplify a filename down to the SDTM list name that should be provided.}
\usage{
make_sdtm_domain_from_filename(filenames = NULL)
}
\arguments{
\item{filenames}{The filenames to simplify.}
}
\description{
Simplify a filename down to the SDTM list name that should be provided.
}
