% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_sdtm.R
\name{make_sdtm_list_name_from_domain}
\alias{make_sdtm_list_name_from_domain}
\title{Extract the domain from a dataset.}
\usage{
make_sdtm_list_name_from_domain(data)
}
\arguments{
\item{data}{The data with a DOMAIN column for extraction}
}
\description{
Extract the domain from a dataset.
}
