% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_sdtm.R
\name{make_sdtm_list_name_from_filename}
\alias{make_sdtm_list_name_from_filename}
\title{Simplify a filename down to the SDTM list name that should be provided.}
\usage{
make_sdtm_list_name_from_filename(filenames)
}
\arguments{
\item{filenames}{The filenames to simplify.}
}
\description{
Simplify a filename down to the SDTM list name that should be provided.
}
