% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_supp.R
\name{merge_supp}
\alias{merge_supp}
\title{Merge a supplementary dataset into a primary dataset}
\usage{
merge_supp(...)
}
\arguments{
\item{...}{Ignored}
}
\description{
This function is defunct. Use `metatools::combine_supp()` instead.
}
