% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_creation.R
\name{new_Rsdtm}
\alias{new_Rsdtm}
\title{Create a new Rsdtm object}
\usage{
new_Rsdtm(data, domain, sdtm_version, ig_name, ig_version)
}
\arguments{
\item{data}{A data.frame or similar object}

\item{domain}{The SDTM domain (as a character scalar)}

\item{ig_name}{The SDTM domain (as a character scalar)}
}
\value{
An Rsdtm object with a class starting c("Rsdtm_<domain>", "Rsdtm"),
  then the original class
}
\description{
Create a new Rsdtm object
}
