% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso8601_regexp.R
\name{pattern_ISO8601_any_date}
\alias{pattern_ISO8601_any_date}
\title{Generate a regular expression matching any ISO8601 date format}
\usage{
pattern_ISO8601_any_date(
  truncated = 0,
  allow_before_year_1583 = FALSE,
  pattern_time = "",
  ...
)
}
\arguments{
\item{truncated}{Should the date be allowed to be truncated?  An integer
indicating the highest required precision (0=day is required, 1=month or week, and
2=year).  A value of 3 will allow an empty string to match.}

\item{allow_before_year_1583}{Should years between 0 and 1582 be allowed
(they are only allowed in ISO 8601 with mutual agreement)}

\item{pattern_time}{A string to add to the day for including time with date
(see \code{pattern_ISO8601_time()})}

\item{...}{Ignored}
}
\description{
Generate a regular expression matching any ISO8601 date format
}
\details{
This matches YYYY-MM-DD (year-month-day), YYYY-Www-d, or YYYY-DDD
  formats. Basic format (without dashes) is not supported.
}
\references{
https://en.wikipedia.org/wiki/ISO_8601
}
\seealso{
Other ISO8601 patterns: 
\code{\link{pattern_ISO8601_any_datetime}()},
\code{\link{pattern_ISO8601_calendar_datetime}()},
\code{\link{pattern_ISO8601_calendar_date}()},
\code{\link{pattern_ISO8601_calendar_year}()},
\code{\link{pattern_ISO8601_ordinal_datetime}()},
\code{\link{pattern_ISO8601_ordinal_date}()},
\code{\link{pattern_ISO8601_timezone}()},
\code{\link{pattern_ISO8601_time}()},
\code{\link{pattern_ISO8601_week_datetime}()},
\code{\link{pattern_ISO8601_week_date}()}
}
\concept{ISO8601 patterns}
