% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso8601_regexp.R
\name{pattern_ISO8601_calendar_datetime}
\alias{pattern_ISO8601_calendar_datetime}
\title{Generate a regular expression matching an ISO8601 calendar date with time}
\usage{
pattern_ISO8601_calendar_datetime(truncated = 0, ...)
}
\arguments{
\item{truncated}{Should the date/time be allowed to be truncated?  An integer
indicating the highest required precision (0=second is required, 1=minute,
2=hour, 3=day, 4=month, 5=year, 6=none).  A value of 6 will allow an empty
string to match.}

\item{...}{Passed to \code{pattern_ISO8601_calendar_date()} and
\code{pattern_ISO8601_time()}}
}
\description{
Generate a regular expression matching an ISO8601 calendar date with time
}
\seealso{
Other ISO8601 patterns: 
\code{\link{pattern_ISO8601_calendar_date}()},
\code{\link{pattern_ISO8601_timezone}()},
\code{\link{pattern_ISO8601_time}()}
}
\concept{ISO8601 patterns}
