% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso8601_regexp.R
\name{pattern_ISO8601_date_builder}
\alias{pattern_ISO8601_date_builder}
\title{Build an ISO8601 date pattern from the year, middle (month or week), day, and
time patterns.}
\usage{
pattern_ISO8601_date_builder(
  truncated,
  pattern_year,
  pattern_middle,
  pattern_day,
  pattern_time,
  allow_before_year_1583
)
}
\arguments{
\item{truncated}{Should the date/time be allowed to be truncated?  An integer
indicating the highest required precision (0=second is required, 1=minute,
2=hour, 3=day, 4=month or week, 5=year, 6=none).  A value of 6 will allow an empty
string to match.}

\item{pattern_year}{The pattern for the ISO8601 year (ignored if
\code{allow_before_year_1583} is given); if null, only the middle and day
part will be returned without a year part.}

\item{pattern_middle}{The pattern for the ISO8601 month or week (or NULL for ordinal dates)}

\item{pattern_day}{The pattern for the ISO8601 day}

\item{pattern_time}{A string to add to the day for including time with date
(see \code{pattern_ISO8601_time()})}

\item{allow_before_year_1583}{Should years between 0 and 1582 be allowed
(they are only allowed in ISO 8601 with mutual agreement)}
}
\description{
Build an ISO8601 date pattern from the year, middle (month or week), day, and
time patterns.
}
\keyword{Internal}
