% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso8601_regexp.R
\name{pattern_ISO8601_time}
\alias{pattern_ISO8601_time}
\title{Generate a regular expression for ISO 8601 times}
\usage{
pattern_ISO8601_time(
  truncated = 0,
  allow_fractional_hours = FALSE,
  allow_fractional_minutes = FALSE,
  allow_fractional_seconds = TRUE,
  fractional_digits = Inf,
  pattern_decimal_mark = c("\\\\.", ","),
  timezone = NA,
  allow_leap_second = FALSE,
  require_T = NA,
  ...
)
}
\arguments{
\item{truncated}{Should the date be allowed to be truncated?  An integer
indicating the highest required precision (0=second is required, 1=minute, and
2=hour).  A value of 3 will allow an empty string to match.}

\item{allow_fractional_hours, allow_fractional_minutes}{Not yet supported}

\item{allow_fractional_seconds}{Should fractional seconds be allowed?}

\item{fractional_digits}{How many digits are allowed in fractional time
units?  \code{0} indicates that they are not allowed, and \code{Inf}
indicates any number are allowed)}

\item{pattern_decimal_mark}{What regular expression patterns should be
allowed for the decimal mark?  Comma and period are both allowed in the
standard and by default.}

\item{timezone}{Should the timezone be optional (NA, default), required
(TRUE), or omitted (FALSE)?}

\item{allow_leap_second}{Should leap seconds (61 seconds in a minute) be
allowed?}

\item{require_T}{Require the "T" at the beginning of the time?  TRUE, FALSE,
or NA (optional)}

\item{...}{Passed to \code{pattern_ISO8601_timezone()}}
}
\description{
Fractional hours and minutes are not allowed in the generated regular
expressions even though they are allowed in the ISO 8601 standard.
}
\details{
Leap seconds occur at 23:59:60Z at the end of the month.  Given timezone
differences which could occur on different hours and at the end of any
15-minute point, \code{allow_leap_seconds} only makes a change to allow for
60 seconds in any minute.
}
\seealso{
Other ISO8601 patterns: 
\code{\link{pattern_ISO8601_calendar_datetime}()},
\code{\link{pattern_ISO8601_calendar_date}()},
\code{\link{pattern_ISO8601_calendar_year}()},
\code{\link{pattern_ISO8601_ordinal_date}()},
\code{\link{pattern_ISO8601_timezone}()},
\code{\link{pattern_ISO8601_week_date}()}
}
\concept{ISO8601 patterns}
