% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso8601_regexp.R
\name{pattern_ISO8601_timezone}
\alias{pattern_ISO8601_timezone}
\title{Generate a regular expression for ISO 8601 timezones.}
\usage{
pattern_ISO8601_timezone()
}
\description{
Generate a regular expression for ISO 8601 timezones.
}
\details{
Offsets are always multiples of 15 minutes, so if the minute part is
  provided, it must be 00, 15, 30, or 45.  The range allowed by this function
  is a bit wider than the current range as of the writing of this function.
  The allowed range is -14:45 to +14:45 while the actual widest range is
  currently -12:00 to +14:00.
  
  According to the ISO standard, zero must always be preceded by a + and
  never a negative.  And, "Z" is allowed.
}
\references{
https://en.wikipedia.org/wiki/UTC_offset
}
\seealso{
Other ISO8601 patterns: 
\code{\link{pattern_ISO8601_calendar_datetime}()},
\code{\link{pattern_ISO8601_calendar_date}()},
\code{\link{pattern_ISO8601_calendar_year}()},
\code{\link{pattern_ISO8601_ordinal_date}()},
\code{\link{pattern_ISO8601_time}()},
\code{\link{pattern_ISO8601_week_date}()}
}
\concept{ISO8601 patterns}
