% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso8601_regexp.R
\name{pattern_ISO8601_week_date}
\alias{pattern_ISO8601_week_date}
\title{Generate a regular expression matching an ISO8601 week date}
\usage{
pattern_ISO8601_week_date(
  truncated = 0,
  allow_before_year_1583 = FALSE,
  pattern_time = "",
  ...
)
}
\arguments{
\item{truncated}{Should the date be allowed to be truncated?  An integer
indicating the highest required precision (0=day is required, 1=month, and
2=year).  A value of 3 will allow an empty string to match.}

\item{allow_before_year_1583}{Should years between 0 and 1582 be allowed
(they are only allowed in ISO 8601 with mutual agreement)}

\item{pattern_time}{A string to add to the day for including time with date
(see \code{pattern_ISO8601_time()})}

\item{...}{Ignored}
}
\description{
Generate a regular expression matching an ISO8601 week date
}
\details{
This matches the general pattern of yyyy-Www-d (year-week of
  year-day of week). Basic format (without dashes) is not supported.
}
\references{
https://en.wikipedia.org/wiki/ISO_8601
}
\seealso{
Other ISO8601 patterns: 
\code{\link{pattern_ISO8601_calendar_datetime}()},
\code{\link{pattern_ISO8601_calendar_date}()},
\code{\link{pattern_ISO8601_calendar_year}()},
\code{\link{pattern_ISO8601_ordinal_date}()},
\code{\link{pattern_ISO8601_timezone}()},
\code{\link{pattern_ISO8601_time}()}
}
\concept{ISO8601 patterns}
