% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_x_VISITDY.R
\name{scale_x_VISITDY}
\alias{scale_x_VISITDY}
\title{A scale for ggplots with SDTM day (DY) numbering}
\usage{
scale_x_VISITDY(...)
}
\arguments{
\item{...}{Passed to \code{ggplot2::scale_x_continuous()}}
}
\description{
The scale will skip over 0, always show 1 (if it is in the range of the
data), provide breaks with preference toward week scales, and provide minor
breaks every day.
}
\examples{
\dontrun{
ggplot(data.frame(x=-7:14, y=-7:14), aes(x=x, y=y, label=x)) +
  scale_x_VISITDY() +
  geom_text()
}
}
\seealso{
Other Output support: 
\code{\link{make_param_label}()}
}
\concept{Output support}
