% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_conversion.R
\name{sdtm_dtc_to_datetime}
\alias{sdtm_dtc_to_datetime}
\alias{sdtm_dtc_to_datetime.list}
\alias{sdtm_dtc_to_datetime.data.frame}
\title{Convert the character representation of the date in an original SDTM dataset
to a POSIXct object.}
\usage{
sdtm_dtc_to_datetime(x, ...)

\method{sdtm_dtc_to_datetime}{list}(x, ...)

\method{sdtm_dtc_to_datetime}{data.frame}(x, date_col_pattern = "DTC$", truncated = 5, ...)
}
\arguments{
\item{x}{The data to convert}

\item{...}{Additional arguments passed to \code{lubridate::ymd_hms}}

\item{date_col_pattern}{A regex to search column names for dates to convert.}

\item{truncated}{Passed to \code{lubridate::ymd_hms}}
}
\value{
The data with the date converted.  Note that all dates will be
  returned as POSIXct objects, so partial dates will appear as the
}
\description{
Convert the character representation of the date in an original SDTM dataset
to a POSIXct object.
}
\seealso{
Other Date management and conversion: 
\code{\link{dateany_to_date}()},
\code{\link{generate_dtc}()},
\code{\link{is_ISO8601_calendar_datetime}()},
\code{\link{is_ISO8601_calendar_date}()},
\code{\link{is_ISO8601_ordinal_datetime}()},
\code{\link{is_ISO8601_timezone}()},
\code{\link{is_ISO8601_time}()},
\code{\link{is_ISO8601_week_datetime}()},
\code{\link{make_dy}()},
\code{\link{pattern_ISO8601_any_datetime}()},
\code{\link{pattern_ISO8601_any_date}()},
\code{\link{pattern_ISO8601_calendar_datetime}()},
\code{\link{pattern_ISO8601_calendar_date}()},
\code{\link{pattern_ISO8601_calendar_year}()},
\code{\link{pattern_ISO8601_ordinal_datetime}()},
\code{\link{pattern_ISO8601_ordinal_date}()},
\code{\link{pattern_ISO8601_timezone}()},
\code{\link{pattern_ISO8601_time}()},
\code{\link{pattern_ISO8601_week_datetime}()},
\code{\link{pattern_ISO8601_week_date}()},
\code{\link{sdtm_first_dose}()},
\code{\link{sdtm_time_actual}()}
}
\concept{Date management and conversion}
