% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm_example_dataverse.R
\docType{data}
\name{sdtm_example_dataverse_ae}
\alias{sdtm_example_dataverse_ae}
\title{Example SDTM Adverse Events Data}
\format{
A data frame with 18 variables and 16 observations:
\describe{
  \item{STUDYID}{Study Identifier, character}
  \item{DOMAIN}{Domain Abbreviation, character}
  \item{USUBJID}{Unique Subject Identifier, character}
  \item{AESEQ}{Sequence Number, numeric}
  \item{AESPID}{Sponsor-Defined Identifier, character}
  \item{AETERM}{Reported Term for the Adverse Event, character}
  \item{AEMODIFY}{Modified Reported Term, character}
  \item{AEDECOD}{Dictionary-Derived Term, character}
  \item{AEBODSYS}{Body System or Organ Class, character}
  \item{AESEV}{Severity/Intensity, character}
  \item{AESER}{Serious Event, character}
  \item{AEACN}{Action Taken with Study Treatment, character}
  \item{AEREL}{Causality, character}
  \item{AESTDTC}{Start Date/Time of Adverse Event, character, ISO8601}
  \item{AEENDTC}{End Date/Time of Adverse Event, character, ISO8601}
  \item{AESTDY}{Study Day of Start of Adverse Event, numeric}
  \item{AEENDY}{Study Day of End of Adverse Event, numeric}
   \item{AEENRF}{End Relative to Reference Period, character}
}
}
\source{
\url{https://doi.org/10.7910/DVN/51B6NK}
}
\usage{
sdtm_example_dataverse_ae
}
\description{
Example SDTM Adverse Events Data
}
\seealso{
Other SDTM example data: 
\code{\link{sdtm_example_dataverse_dm}},
\code{\link{sdtm_example_dataverse_eg}},
\code{\link{sdtm_example_dataverse_lb}},
\code{\link{sdtm_example_dataverse_mh}},
\code{\link{sdtm_example_dataverse_pe}},
\code{\link{sdtm_example_dataverse_vs}}
}
\concept{SDTM example data}
\keyword{datasets}
