% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm_example_dataverse.R
\docType{data}
\name{sdtm_example_dataverse_dm}
\alias{sdtm_example_dataverse_dm}
\title{Example SDTM Demographics Data}
\format{
A data frame with 16 variables and 5 observations:
\describe{
 \item{STUDYID }{Study Identifier, character}
 \item{DOMAIN }{Domain Abbreviation, character}
 \item{USUBJID }{Unique Subject Identifier, character}
 \item{SUBJID }{Subject Identifier for the Study, character}
 \item{RFSTDTC }{Subject Reference Start Date/Time, character, ISO 8601}
 \item{RFENDTC }{Subject Reference End Date/Time, character, ISO 8601}
 \item{SITEID }{Study Site Identifier, character}
 \item{BRTHDTC }{Date/Time of Birth, character, ISO 8601}
 \item{AGE }{Age, numeric}
 \item{AGEU }{Age Units, character}
 \item{SEX }{Sex, character}
 \item{RACE }{Race , character}
 \item{ETHNIC }{Ethnicity, character}
 \item{ARMCD }{Planned Arm Code, character}
 \item{ARM }{Description of Planned Arm, character}
 \item{COUNTRY }{Country, character}
}
}
\source{
\url{https://doi.org/10.7910/DVN/51B6NK}
}
\usage{
sdtm_example_dataverse_dm
}
\description{
Example SDTM Demographics Data
}
\seealso{
Other SDTM example data: 
\code{\link{sdtm_example_dataverse_ae}},
\code{\link{sdtm_example_dataverse_eg}},
\code{\link{sdtm_example_dataverse_lb}},
\code{\link{sdtm_example_dataverse_mh}},
\code{\link{sdtm_example_dataverse_pe}},
\code{\link{sdtm_example_dataverse_vs}},
\code{\link{sdtm_example_theoph_ex}},
\code{\link{sdtm_example_theoph_pc}}
}
\concept{SDTM example data}
\keyword{datasets}
