% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm_example_dataverse.R
\docType{data}
\name{sdtm_example_dataverse_eg}
\alias{sdtm_example_dataverse_eg}
\title{Example SDTM ECG Test Results Data}
\format{
A data frame with 21 variables and 56 observations:
\describe{
 \item{STUDYID}{Study Identifier, character}
 \item{DOMAIN}{Domain Abbreviation, character}
 \item{USUBJID}{Unique Subject Identifier, character}
 \item{EGSEQ}{Sequence Number, numeric}
 \item{EGSPID}{Sponsor-Defined Identifier, character}
 \item{EGTESTCD}{ECG Test or Examination Short Name, character}
 \item{EGTEST}{ECG Test or Examination Name, character}
 \item{EGPOS}{ECG Position of Subject, character}
 \item{EGORRES}{Result or Finding in Original Units, character}
 \item{EGORRESU}{Original Units, character}
 \item{EGSTRESC}{Character Result/Finding in Std Format, character}
 \item{EGSTRESN}{Numeric Result/Finding in Standard Units, numeric}
 \item{EGSTRESU}{Standard Units, character}
 \item{EGBLFL}{Baseline Flag, character}
 \item{EGDRVFL}{Derived Flag, character}
 \item{EGEVAL}{Evaluator, character}
 \item{VISITNUM}{Visit Number, numeric}
 \item{VISIT}{Visit Name, character}
 \item{VISITDY}{Planned Study Day of Visit, numeric}
 \item{EGDTC}{Date/Time of ECG, character, ISO 8601}
 \item{EGDY}{Study Day of ECG, numeric}
}
}
\source{
\url{https://doi.org/10.7910/DVN/51B6NK}
}
\usage{
sdtm_example_dataverse_eg
}
\description{
Example SDTM ECG Test Results Data
}
\seealso{
Other SDTM example data: 
\code{\link{sdtm_example_dataverse_ae}},
\code{\link{sdtm_example_dataverse_dm}},
\code{\link{sdtm_example_dataverse_lb}},
\code{\link{sdtm_example_dataverse_mh}},
\code{\link{sdtm_example_dataverse_pe}},
\code{\link{sdtm_example_dataverse_vs}},
\code{\link{sdtm_example_theoph_ex}},
\code{\link{sdtm_example_theoph_pc}}
}
\concept{SDTM example data}
\keyword{datasets}
