% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm_example_dataverse.R
\docType{data}
\name{sdtm_example_dataverse_lb}
\alias{sdtm_example_dataverse_lb}
\title{Example SDTM Laboratory Test Results Data}
\format{
A data frame with 29 variables and 83 observations:
\describe{
 \item{STUDYID}{Study Identifier, character}
 \item{DOMAIN}{Domain Abbreviation, character}
 \item{USUBJID}{Unique Subject Identifier, character}
 \item{LBSEQ}{Sequence Number, numeric}
 \item{LBREFID}{Specimen ID, character}
 \item{LBTESTCD}{Lab Test or Examination Short Name, character}
 \item{LBTEST}{Lab Test or Examination Name, character}
 \item{LBCAT}{Category for Lab Test, character}
 \item{LBSCAT}{Subcategory for Lab Test, character}
 \item{LBORRES}{Result or Finding in Original Units, character}
 \item{LBORRESU}{Original Units, character}
 \item{LBORNRLO}{Reference Range Lower Limit in Orig Unit, character}
 \item{LBORNRHI}{Reference Range Upper Limit in Orig Unit, character}
 \item{LBSTRESC}{Character Result/Finding in Std Format, character}
 \item{LBSTRESN}{Numeric Result/Finding in Standard Units, numeric}
 \item{LBSTRESU}{Standard Units, character}
 \item{LBSTNRLO}{Reference Range Lower Limit-Std Units, numeric}
 \item{LBSTNRHI}{Reference Range Upper Limit-Std Units, numeric}
 \item{LBSTNRC}{Reference Range for Char Rslt-Std Units, character}
 \item{LBNRIND}{Reference Range Indicator, character}
 \item{LBSPEC}{Specimen Type, character}
 \item{LBMETHOD}{Method of Test or Examination, character}
 \item{LBBLFL}{Baseline Flag, character}
 \item{LBFAST}{Fasting Status, character}
 \item{VISITNUM}{Visit Number, numeric}
 \item{VISIT}{Visit Name, character}
 \item{VISITDY}{Planned Study Day of Visit, numeric}
 \item{LBDTC}{Date/time of Specimen Collection, character, ISO 8601}
 \item{LBDY}{Study Day of Specimen Collection, numeric}
}
}
\source{
\url{https://doi.org/10.7910/DVN/51B6NK}
}
\usage{
sdtm_example_dataverse_lb
}
\description{
Example SDTM Laboratory Test Results Data
}
\seealso{
Other SDTM example data: 
\code{\link{sdtm_example_dataverse_ae}},
\code{\link{sdtm_example_dataverse_dm}},
\code{\link{sdtm_example_dataverse_eg}},
\code{\link{sdtm_example_dataverse_mh}},
\code{\link{sdtm_example_dataverse_pe}},
\code{\link{sdtm_example_dataverse_vs}},
\code{\link{sdtm_example_theoph_ex}},
\code{\link{sdtm_example_theoph_pc}}
}
\concept{SDTM example data}
\keyword{datasets}
