% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm_example_dataverse.R
\docType{data}
\name{sdtm_example_dataverse_mh}
\alias{sdtm_example_dataverse_mh}
\title{Example SDTM Medical History Data}
\format{
A data frame with 12 variables and 18 observations:
\describe{
 \item{STUDYID}{Study Identifier, character}
 \item{DOMAIN}{Domain Abbreviation, character}
 \item{USUBJID}{Unique Subject Identifier, character}
 \item{MHSEQ}{Sequence Number, numeric}
 \item{MHTERM}{Reported Term for the Medical History, character}
 \item{MHCAT}{Category for Medical History, character}
 \item{MHPRESP}{Medical History Event Pre- Specified, character}
 \item{MHOCCUR}{Medical History Occurrence, character}
 \item{MHBODSYS}{Body System or Organ Class, character}
 \item{MHDTC}{Date/Time of History Collection, character, ISO 8601}
 \item{MHSTDTC}{Start Date/Time of Medical History Event, character, ISO 8601}
 \item{MHENRF}{End Relative to Reference Period, character}
 }
}
\source{
\url{https://doi.org/10.7910/DVN/51B6NK}
}
\usage{
sdtm_example_dataverse_mh
}
\description{
Example SDTM Medical History Data
}
\seealso{
Other SDTM example data: 
\code{\link{sdtm_example_dataverse_ae}},
\code{\link{sdtm_example_dataverse_dm}},
\code{\link{sdtm_example_dataverse_eg}},
\code{\link{sdtm_example_dataverse_lb}},
\code{\link{sdtm_example_dataverse_pe}},
\code{\link{sdtm_example_dataverse_vs}},
\code{\link{sdtm_example_theoph_ex}},
\code{\link{sdtm_example_theoph_pc}}
}
\concept{SDTM example data}
\keyword{datasets}
