% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm_example_dataverse.R
\docType{data}
\name{sdtm_example_dataverse_pe}
\alias{sdtm_example_dataverse_pe}
\title{Example SDTM Physical Examination Data}
\format{
A data frame with 13 variables and 65 observations:
\describe{
 \item{STUDYID}{Study Identifier, character}
 \item{DOMAIN}{Domain Abbreviation, character}
 \item{USUBJID}{Unique Subject Identifier, character}
 \item{PESEQ}{Sequence Number, numeric}
 \item{PETESTCD}{Body System Examined Short Name, character}
 \item{PETEST}{Body System Examined, character}
 \item{PEORRES}{Verbatim Examination Finding, character}
 \item{PESTRESC}{Character Result/Finding in Standard Format, character}
 \item{VISITNUM}{Visit Number, numeric}
 \item{VISIT}{Visit Name, character}
 \item{VISITDY}{Planned Study Day of Visit, numeric}
 \item{PEDTC}{Date/Time of Examination, character, ISO 8601}
 \item{PEDY}{Study Day of Examination, numeric}
 }
}
\source{
\url{https://doi.org/10.7910/DVN/51B6NK}
}
\usage{
sdtm_example_dataverse_pe
}
\description{
Example SDTM Physical Examination Data
}
\seealso{
Other SDTM example data: 
\code{\link{sdtm_example_dataverse_ae}},
\code{\link{sdtm_example_dataverse_dm}},
\code{\link{sdtm_example_dataverse_eg}},
\code{\link{sdtm_example_dataverse_lb}},
\code{\link{sdtm_example_dataverse_mh}},
\code{\link{sdtm_example_dataverse_vs}}
}
\concept{SDTM example data}
\keyword{datasets}
