% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm_example_dataverse.R
\docType{data}
\name{sdtm_example_dataverse_vs}
\alias{sdtm_example_dataverse_vs}
\title{Example SDTM Vital Signs Data}
\format{
A data frame with 18 variables and 72 observations:
\describe{
 \item{STUDYID}{Study Identifier, character}
 \item{DOMAIN}{Domain Abbreviation, character}
 \item{USUBJID}{Unique Subject Identifier, character}
 \item{VSSEQ}{Sequence Number, numeric}
 \item{VSTESTCD}{Vital Signs Test Short Name, character}
 \item{VSTEST}{Vital Signs Test Name, character}
 \item{VSPOS}{Vital Signs Position of Subject, character}
 \item{VSORRES}{Result or Finding in Original Units, character}
 \item{VSORRESU}{Original Units, character}
 \item{VSSTRESC}{Character Result/Finding in Std Format, character}
 \item{VSSTRESN}{Numeric Result/Finding in Standard Units, numeric}
 \item{VSSTRESU}{Standard Units, character}
 \item{VSBLFL}{Baseline Flag, character}
 \item{VISITNUM}{Visit Number, numeric}
 \item{VISIT}{Visit Name, character}
 \item{VISITDY}{Planned Study Day of Visit, numeric}
 \item{VSDTC}{Date/Time of Measurements, character, ISO 8601}
 \item{VSDY}{Study Day of Vital Signs, numeric}
 }
}
\source{
\url{https://doi.org/10.7910/DVN/51B6NK}
}
\usage{
sdtm_example_dataverse_vs
}
\description{
Example SDTM Vital Signs Data
}
\seealso{
Other SDTM example data: 
\code{\link{sdtm_example_dataverse_ae}},
\code{\link{sdtm_example_dataverse_dm}},
\code{\link{sdtm_example_dataverse_eg}},
\code{\link{sdtm_example_dataverse_lb}},
\code{\link{sdtm_example_dataverse_mh}},
\code{\link{sdtm_example_dataverse_pe}}
}
\concept{SDTM example data}
\keyword{datasets}
