% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm_example_theoph.R
\docType{data}
\name{sdtm_example_theoph_ex}
\alias{sdtm_example_theoph_ex}
\title{Example SDTM Exposure Data}
\format{
A data frame with 9 variables and 12 observations:
\describe{
 \item{STUDYID}{Study Identifier, character}
 \item{DOMAIN}{Domain Abbreviation, character}
 \item{SUBJID}{Subject Identifier for the Study, character}
 \item{USUBJID}{Unique Subject Identifier, character}
 \item{EXSEQ}{Sequence Number, numeric}
 \item{EXTRT}{Name of Treatment, character}
 \item{EXDOSE}{Dose, Numeric}
 \item{EXDOSU}{Dose Units, character}
 \item{EXROUTE}{Route of Administration, character}
 \item{EXSTDTC}{Start Date/Time of Treatment, character, ISO 8601}
}
}
\usage{
sdtm_example_theoph_ex
}
\description{
Example SDTM Exposure Data
}
\seealso{
Other SDTM example data: 
\code{\link{sdtm_example_dataverse_ae}},
\code{\link{sdtm_example_dataverse_dm}},
\code{\link{sdtm_example_dataverse_eg}},
\code{\link{sdtm_example_dataverse_lb}},
\code{\link{sdtm_example_dataverse_mh}},
\code{\link{sdtm_example_dataverse_pe}},
\code{\link{sdtm_example_dataverse_vs}},
\code{\link{sdtm_example_theoph_pc}}
}
\concept{SDTM example data}
\keyword{datasets}
