% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm_example_theoph.R
\docType{data}
\name{sdtm_example_theoph_pc}
\alias{sdtm_example_theoph_pc}
\title{Example SDTM Pharmacokinetics Concentration Data}
\format{
A data frame with 15 variables and 132 observations:
\describe{
 \item{STUDYID}{Study Identifier, character}
 \item{DOMAIN}{Domain Abbreviation, character}
 \item{SUBJID}{Subject Identifier for the Study, character}
 \item{USUBJID}{Unique Subject Identifier, character}
 \item{PCSEQ}{Sequence Number, numeric}
 \item{PCTESTCD}{Pharmacokinetic Test Short Name, character}
 \item{PCTEST}{Pharmacokinetic Test Name, character}
 \item{PCORRES}{Result or Finding in Original Units, character}
 \item{PCORRESU}{Original Units, character}
 \item{PCSTRESC}{Character Result/Finding in Standard Format, character}
 \item{PCSTRESN}{Numeric Result/Finding in Standard Units, numeric}
 \item{PCSTRESU}{Standard Units, character}
 \item{PCDTC}{Date/Time of Specimen Collection, character, ISO 8601}
 \item{PCTPT}{Planned Time Point Name, character}
 \item{PCTPTNUM}{Planned Time Point Number, numeric}
 }
}
\usage{
sdtm_example_theoph_pc
}
\description{
Example SDTM Pharmacokinetics Concentration Data
}
\seealso{
Other SDTM example data: 
\code{\link{sdtm_example_dataverse_ae}},
\code{\link{sdtm_example_dataverse_dm}},
\code{\link{sdtm_example_dataverse_eg}},
\code{\link{sdtm_example_dataverse_lb}},
\code{\link{sdtm_example_dataverse_mh}},
\code{\link{sdtm_example_dataverse_pe}},
\code{\link{sdtm_example_dataverse_vs}},
\code{\link{sdtm_example_theoph_ex}}
}
\concept{SDTM example data}
\keyword{datasets}
