% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify_sdtm_names.R
\name{simplify_sdtm_names}
\alias{simplify_sdtm_names}
\title{Simplify the names of an SDTM data.frame to remove the DOMAIN from name
prefixes.}
\usage{
simplify_sdtm_names(x)
}
\arguments{
\item{x}{A data.frame of SDTM data from a single domain.}
}
\value{
\code{x} with the DOMAIN removed from names.
}
\description{
Simplify the names of an SDTM data.frame to remove the DOMAIN from name
prefixes.
}
\details{
WARNING: Currently, this is implemented by looking at the domain and
  stripping those characters from the beginning of any column names.  If
  there are column names that incidentally start with the same two
  characters, the initial two characters will be removed in error (please
  report that as a bug).
}
