% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_ids.R
\name{standardize_sdtm_id}
\alias{standardize_sdtm_id}
\title{Standardize the SDTM ID columns (STUDYID, USUBJID, and SUBJID)}
\usage{
standardize_sdtm_id(
  data,
  id,
  studyid,
  sep_usubjid = "-",
  allow_missing_id = FALSE
)
}
\arguments{
\item{data}{data.frame or similar to have standard IDs added}

\item{id}{a data.frame or list of data.frames to merge other identifiers into
the data}

\item{studyid}{a character scalar indicating the study number}

\item{sep_usubjid}{a character scalar separating the STUDYID from the SUBJID}

\item{allow_missing_id}{Is it acceptable for some USUBJID and SUBJID to be
missing?}
}
\value{
A data.frame with columns for STUDYID, USUBJID, and SUBJID and other
  identifier columns removed.
}
\description{
Standardize the SDTM ID columns (STUDYID, USUBJID, and SUBJID)
}
