% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_methods.R
\name{[[.Rsdtm}
\alias{[[.Rsdtm}
\title{Extract a column from an Rsdtm object}
\usage{
\method{[[}{Rsdtm}(data, i)
}
\arguments{
\item{data}{An Rsdtm object}

\item{i}{The column name.  It may be a domain-agnostic column name.}
}
\value{
The column
}
\description{
Extract a column from an Rsdtm object
}
\examples{
object[["LBTEST"]]
object[["--TEST"]]
}
