% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_supp.R
\name{supp_reformat}
\alias{supp_reformat}
\title{Reformat a --SUPP SDTM domain into a list of data.frames ready for merging
into the primary domain.}
\usage{
supp_reformat(x, auto_convert = TRUE)
}
\arguments{
\item{x}{a --SUPP SDTM domain object}

\item{auto_convert}{should the data be automatically converted?}
}
\value{
A list with length the same as \code{unique(x$IDVAR)} with
  data.frames ready for merging into the primary dataset.
}
\description{
Reformat a --SUPP SDTM domain into a list of data.frames ready for merging
into the primary domain.
}
\seealso{
\code{\link{merge_supp}}
}
