% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{valid_versions}
\alias{valid_versions}
\title{A data.frame with 4 columns and one row per valid SDTM version/IG/IG version
combination}
\format{
\itemize{
  \item{SDTM}{The SDTM version}
  \item{IG_NAME}{The implementation guide (IG) name being used}
  \item{IG_VERSION}{The implemetation guide (IG) version}
  \item{archived}{The SDTM version/IG pair has been archived (and should not
    be used for current studies)}
}
}
\usage{
valid_versions
}
\description{
A data.frame with 4 columns and one row per valid SDTM version/IG/IG version
combination
}
\keyword{datasets}
