% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_column.r
\name{validate_column_class}
\alias{validate_column_class}
\title{Check to see if columns have the expected class (numeric or character), if not give a warning and convert to one or the other. Also, check to see if columns contain NA, give an error if they do.}
\usage{
validate_column_class(data, domain_spec)
}
\arguments{
\item{data}{The dataset whose columns to check.}

\item{domain_spec}{The domain specification ("CO", "DM", "SE", "SM", "SV") for checking columns.}
}
\value{
If columns are numeric but should be character, give a warning and indicate which are those columns. If columns are character but should be numeric, give a warning and indicate those columns. If columns have NA, give a warning and indicate which are those columns.
}
\description{
Check to see if columns have the expected class (numeric or character), if not give a warning and convert to one or the other. Also, check to see if columns contain NA, give an error if they do.
}
