% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_column.r
\name{validate_column_names}
\alias{validate_column_names}
\title{Check if dataset is consistent with SDTMIG specifications.}
\usage{
validate_column_names(data, sdtmig_version, domain_name)
}
\arguments{
\item{data}{The dataset whose columns to check.}

\item{sdtmig_version}{The version of SDTMIG (e.g. "SDTMIG3.3").}

\item{domain_name}{The domain specification ("CO", "DM", "SE", "SM", "SV") for checking columns.}
}
\value{
If dataset is not consistent with SDTMIG specifications, give a warning or error and convert class-misspecified columns to specified class.
}
\description{
Check if dataset is consistent with SDTMIG specifications.
}
