% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_column.r
\name{validate_ctrl_terminology}
\alias{validate_ctrl_terminology}
\title{Check to see if columns have controlled terminology, give an error if columns do not have controlled terminology.}
\usage{
validate_ctrl_terminology(data, domain_spec)
}
\arguments{
\item{data}{The dataset whose columns to check.}

\item{domain_spec}{The domain specification ("CO", "DM", "SE", "SM", "SV") for checking columns.}
}
\value{
If columns contain do not have controlled terminology, give an error and indicate which are those columns.
}
\description{
Check to see if columns have controlled terminology, give an error if columns do not have controlled terminology.
}
