% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_column.r
\name{validate_expected_column_names}
\alias{validate_expected_column_names}
\title{Check if expected columns exist, give a warning if they do not exist}
\usage{
validate_expected_column_names(data, domain_spec)
}
\arguments{
\item{data}{The dataset whose columns to check.}

\item{domain_spec}{The domain specification ("CO", "DM", "SE", "SM", "SV") for checking columns.}
}
\value{
If expected columns are missing, give an warning and indicate which expected columns are missing.
}
\description{
Check if expected columns exist, give a warning if they do not exist
}
