% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_column.r
\name{validate_extra_column_names}
\alias{validate_extra_column_names}
\title{Check if extra columns exist, give an warning if they do exist}
\usage{
validate_extra_column_names(data, domain_spec)
}
\arguments{
\item{data}{The dataset whose columns to check.}

\item{domain_spec}{The domain specification ("CO", "DM", "SE", "SM", "SV") for checking columns.}
}
\value{
If extra columns exist, give an warning and indicate which extra columns exist.
}
\description{
Check if extra columns exist, give an warning if they do exist
}
