% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_column.r
\name{validate_iso}
\alias{validate_iso}
\title{If the dataset has COUNTRY column, check to see if the COUNTRY code is consistent with ISO 3166-1 Alpha-3.}
\usage{
validate_iso(data, domain_spec)
}
\arguments{
\item{data}{The dataset whose columns to check.}

\item{domain_spec}{The domain specification ("CO", "DM", "SE", "SM", "SV") for checking columns.}
}
\value{
Give an error if the COUNTRY code is not consistent with ISO 3166-1 Alpha-3.
}
\description{
If the dataset has COUNTRY column, check to see if the COUNTRY code is consistent with ISO 3166-1 Alpha-3.
}
