% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_column.r
\name{validate_permitted_column_names}
\alias{validate_permitted_column_names}
\title{Check if permitted columns exist, no message either way}
\usage{
validate_permitted_column_names(data, domain_spec)
}
\arguments{
\item{data}{The dataset whose columns to check.}

\item{domain_spec}{The domain specification ("CO", "DM", "SE", "SM", "SV") for checking columns.}
}
\value{
No message regardless of whether permitted columns are missing.
}
\description{
Check if permitted columns exist, no message either way
}
