% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_column.r
\name{validate_required_column_names}
\alias{validate_required_column_names}
\title{Check if required columns exist, give an error if they do not exist.}
\usage{
validate_required_column_names(data, domain_spec)
}
\arguments{
\item{data}{The dataset whose columns to check.}

\item{domain_spec}{The domain specification ("CO", "DM", "SE", "SM", "SV") for checking columns.}
}
\value{
If required columns are missing, give an error and indicate which required columns are missing.
}
\description{
Check if required columns exist, give an error if they do not exist.
}
