% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_sdtm.R
\name{append_no_duplicate_names}
\alias{append_no_duplicate_names}
\title{Add elements to a vector, only if names are not duplicated in \code{x} and
\code{values}}
\usage{
append_no_duplicate_names(x, values, after = length(x),
  ignore_blank = TRUE, method = stop)
}
\arguments{
\item{x, values, after}{See \code{base::append}}

\item{ignore_blank}{Ignore blank names? (Blank names \code{""} occur when an
unnamed vector is appeneded to a named vector.)}

\item{method}{Function to notify the user if a duplicate is provided?  It is
called with a message indicating the name of the duplicate. Typically this
will be \code{stop} or \code{warning}.}
}
\value{
See \code{base::append}
}
\description{
Add elements to a vector, only if names are not duplicated in \code{x} and
\code{values}
}
