% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso8601_regexp.R
\name{is_ISO8601_timezone}
\alias{is_ISO8601_timezone}
\title{Determine if a string is an ISO 8601 timezone}
\usage{
is_ISO8601_timezone(x)
}
\arguments{
\item{x}{A vector of character strings to test}
}
\description{
Determine if a string is an ISO 8601 timezone
}
\seealso{
Other ISO8601 String checking: 
\code{\link{is_ISO8601_calendar_datetime}()},
\code{\link{is_ISO8601_calendar_date}()},
\code{\link{is_ISO8601_time}()}
}
\concept{ISO8601 String checking}
