% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_supp.R
\name{merge_supp}
\alias{merge_supp}
\title{Merge a supplementary dataset into a primary dataset}
\usage{
merge_supp(primary, supplementary, remove_attributes = c("label"))
}
\arguments{
\item{primary}{the data.frame of the primary dataset}

\item{supplementary}{the data.frame of the supplementary dataset}

\item{remove_attributes}{If \code{TRUE}, remove all attributes from all
columns (this will break many classes); if \code{FALSE}, remove no
attributes from any columns; if a character, remove those attributes from
the columns.}
}
\value{
\code{primary} merged with \code{supplementary} where new column
  names come from the \code{QNAM} column in \code{supplementary}
}
\description{
Merge a supplementary dataset into a primary dataset
}
\seealso{
\code{\link{supp_reformat}}
}
