% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_conversion.R
\name{sdtm_dtc_to_datetime}
\alias{sdtm_dtc_to_datetime}
\alias{sdtm_dtc_to_datetime.list}
\alias{sdtm_dtc_to_datetime.data.frame}
\title{Convert the character representation of the date in an original SDTM dataset
to a POSIXct object.}
\usage{
sdtm_dtc_to_datetime(x, ...)

\method{sdtm_dtc_to_datetime}{list}(x, ...)

\method{sdtm_dtc_to_datetime}{data.frame}(x, date_col_pattern = "DTC$",
  truncated = 5, ...)
}
\arguments{
\item{x}{The data to convert}

\item{...}{Additional arguments passed to \code{lubridate::ymd_hms}}

\item{date_col_pattern}{A regex to search column names for dates to convert.}

\item{truncated}{Passed to \code{lubridate::ymd_hms}}
}
\value{
The data with the date converted.  Note that all dates will be
  returned as POSIXct objects, so partial dates will appear as the
}
\description{
Convert the character representation of the date in an original SDTM dataset
to a POSIXct object.
}
