% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_actual.R
\name{sdtm_time_actual}
\alias{sdtm_time_actual}
\alias{sdtm_time_actual.list}
\alias{sdtm_time_actual.data.frame}
\title{Generate actual times from dates and times}
\usage{
sdtm_time_actual(x, ...)

\method{sdtm_time_actual}{list}(x, ...)

\method{sdtm_time_actual}{data.frame}(
  x,
  ref_col = "DTC_first_dose",
  dtc_col_pattern = "DTC$",
  units = "hours",
  tz = NULL,
  ...
)
}
\arguments{
\item{x}{An SDTM data set}

\item{...}{Passed to methods.}

\item{ref_col}{The character name of the column to use as a reference time
point.}

\item{dtc_col_pattern}{The regex to use to select columns to use for actual
time generation.  The regex will also be used to modify the name.}

\item{units}{The units to report the actual time in (passed to \code{as.numeric}).}

\item{tz}{The time zone for difftime (default is used if \code{is.null(tz)}).}
}
\value{
\code{x} with new columns matching \code{dtc_col_pattern} named
  \code{gsub(pattern=dtc_col_pattern, replacement=paste0("actual_", units),
  names(x))} The new columns will have the actual time in numerical values
  matching the \code{units}.
}
\description{
Generate actual times from dates and times
}
\seealso{
Other Date management and conversion: 
\code{\link{dateany_to_date}()},
\code{\link{generate_dtc}()},
\code{\link{is_ISO8601_calendar_datetime}()},
\code{\link{is_ISO8601_calendar_date}()},
\code{\link{is_ISO8601_ordinal_datetime}()},
\code{\link{is_ISO8601_timezone}()},
\code{\link{is_ISO8601_time}()},
\code{\link{is_ISO8601_week_datetime}()},
\code{\link{make_dy}()},
\code{\link{pattern_ISO8601_any_datetime}()},
\code{\link{pattern_ISO8601_any_date}()},
\code{\link{pattern_ISO8601_calendar_datetime}()},
\code{\link{pattern_ISO8601_calendar_date}()},
\code{\link{pattern_ISO8601_calendar_year}()},
\code{\link{pattern_ISO8601_ordinal_datetime}()},
\code{\link{pattern_ISO8601_ordinal_date}()},
\code{\link{pattern_ISO8601_timezone}()},
\code{\link{pattern_ISO8601_time}()},
\code{\link{pattern_ISO8601_week_datetime}()},
\code{\link{pattern_ISO8601_week_date}()},
\code{\link{sdtm_dtc_to_datetime}()},
\code{\link{sdtm_first_dose}()}
}
\concept{Date management and conversion}
